# tee-broker/broker-daemon/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VerdantForged TEE Broker API",
    "version": "0.2.0",
    "summary": "REST API for the VerdantForged TEE broker (eu-west-1).",
    "description": "Submit attested LLM jobs to a TEE-backed worker, poll status, fetch encrypted artifacts, and register new skills. The broker is a cheap always-on control plane (t3.small) that holds the queue, per-job LLM tokens, and Stripe payment lifecycle. The worker (m6a.xlarge, AMD SEV-SNP) is launched on demand and terminated after a configurable idle buffer. The real LLM API key never enters the worker — the worker receives per-job ephemeral tokens and calls the broker's /v1/llm/chat/completions proxy. **Live broker:** https://verdant.codepilots.co.uk",
    "contact": {
      "name": "VerdantForged",
      "url": "https://verdantforged.pages.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://verdant.codepilots.co.uk",
      "description": "Live broker (eu-west-1)"
    }
  ],
...
  "paths": {
...
    "/v1/skills": {
      "get": {
        "tags": ["Skills"],
        "summary": "List registered skills",
        "description": "Returns every distinct skill name with its latest version. Built-in stubs (`code-review`, `summarize`, `photo-glow-up`) are advertised via `/v1/discover` and do NOT appear here unless explicitly registered. No auth.",
        "operationId": "listSkills",
        "responses": {
          "200": {
            "description": "Skill catalog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["skills"],
                  "properties": {
                    "skills": {"type": "array", "items": {"$ref": "#/components/schemas/Skill"}}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Skills"],
        "summary": "Register a new skill",
        "description": "Registers a new skill (or a new version of an existing one). Exactly one of `prompt_template` or `wasm_ref` must be provided. Resource limits are bounded above the built-in stubs (default 256 MB / 60 s / 10M fuel) and capped below 4 GB / 600 s / 100M fuel to prevent a compromised publisher from exhausting the broker. **Auth (required):** `Authorization: Bearer <BROKER_SKILLS_API_KEY>`. If the server's key is unset the endpoint returns 503.",
        "operationId": "registerSkill",
        "security": [{"SkillsApiKey": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/SkillManifest"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Skill registered",
            "headers": {
              "Location": {"description": "Canonical URL for the registered skill", "schema": {"type": "string"}}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Skill"}
              }
            }
          },
          "400": {"description": "Validation failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"description": "Missing or invalid Authorization header", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": {"description": "Skill already registered at this version", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "503": {"description": "Skill registration is disabled (BROKER_SKILLS_API_KEY not configured on server)", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/v1/skills/{ref}": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get a specific skill",
        "description": "Returns a single skill. `ref` is either `name` (latest version) or `name@version`. No auth.",
        "operationId": "getSkill",
        "parameters": [
          {"name": "ref", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Skill record", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Skill"}}}},
          "404": {"description": "Skill not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SkillsApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-side `BROKER_SKILLS_API_KEY`. Used only for `POST /v1/skills`."
      },
      "LlmToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal per-job worker token. It is placed only in the EFS worker envelope and is never returned to clients."
      },
      "JobAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "One-time client credential returned only by the initial successful POST /v1/jobs response. Stored server-side as a SHA-256 digest."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {"type": "string"},
          "code": {"type": "string", "description": "Stable machine-readable error code (when applicable)."}
        }
      },
...
      "SkillManifest": {
        "type": "object",
        "required": ["name", "description", "wasm_manifest_hash", "entry_point"],
        "properties": {
          "name": {"type": "string", "pattern": "^[a-z0-9][a-z0-9-]{2,63}$", "description": "Lowercase slug, 3-64 chars, no leading/trailing dash."},
          "version": {"type": "string", "pattern": "^[0-9]+\\.[0-9]+\\.[0-9]+$", "default": "0.1.0"},
          "description": {"type": "string", "minLength": 1, "maxLength": 512},
          "wasm_manifest_hash": {"type": "string", "pattern": "^[0-9a-fA-F]{64}$", "description": "SHA-256 of the WASM binary."},
          "entry_point": {"type": "string", "pattern": "^[A-Za-z0-9_.]{1,128}$"},
          "prompt_template": {"type": "string", "maxLength": 32768, "description": "Provide exactly one of `prompt_template` or `wasm_ref`."},
          "wasm_ref": {"type": "object", "description": "Provide exactly one of `prompt_template` or `wasm_ref`.", "required": ["uri", "size_bytes"], "properties": {"uri": {"type": "string", "maxLength": 512}, "size_bytes": {"type": "integer", "minimum": 1, "maximum": 52428800}}},
          "resource_limits": {"type": "object", "properties": {
            "max_fuel": {"type": "integer", "minimum": 1, "maximum": 100000000, "default": 10000000},
            "max_duration_ms": {"type": "integer", "minimum": 100, "maximum": 600000, "default": 60000},
            "max_memory_mb": {"type": "integer", "minimum": 1, "maximum": 4096, "default": 256}
          }},
          "input_schema": {"type": "object"},
          "output_schema": {"type": "object"}
        }
      },
      "Skill": {"$ref": "#/components/schemas/SkillManifest"},
...
    }
  }
}